+++ test/Tuple.ts
@@ -9,12 +9,14 @@ import {
   create,
   mapBoth,
   fanout,
+  getEq,
 } from "../src/Tuple"
 import { increment } from "../src/Number"
 import { constant, flow, identity, pipe } from "fp-ts/function"
 import { bimap, swap } from "fp-ts/Tuple"
 import * as O from "fp-ts/Option"
 import fc from "fast-check"
+import * as Str from "fp-ts/string"
 
 const nonMaxNumber = fc.integer({ max: Number.MAX_SAFE_INTEGER - 1 })
 
@@ -171,6 +173,28 @@ describe("Tuple", () => {
         ),
       )
       expect(f)
+    })
+  })
+
+  describe("getEq", () => {
+    it("checks both values in terms of AND", () => {
+      const { equals: f } = getEq(Str.Eq)(Str.Eq)
+
+      expect(f(["foo", "foo"], ["foo", "foo"])).toBe(true)
+      expect(f(["foo", "foo"], ["bar", "foo"])).toBe(false)
+      expect(f(["foo", "foo"], ["foo", "bar"])).toBe(false)
+      expect(f(["foo", "bar"], ["baz", "oof"])).toBe(false)
+    })
+
+    it("checks second component lazily", () => {
+      const { equals: f } = getEq(Str.Eq)({
+        equals: () => {
+          // eslint-disable-next-line functional/no-throw-statements
+          throw "evaluated second component"
+        },
+      })
+
+      expect(f(["foo", "foo"], ["bar", "foo"])).toBe(false)
     })
   })
 })